+++ .editorconfig
@@ -3,8 +3,8 @@ root = true
 [*]
 charset = utf-8
 end_of_line = lf
-indent_size = 2
 indent_style = space
+indent_size = 2
 insert_final_newline = true
 tab_width = 2
 trim_trailing_whitespace = true
