+++ .env
@@ -1,7 +1,7 @@
 ENVIRONMENT=development
 HOST=0.0.0.0
 PORT=4000
-URL=localhost
 TITLE=React
 DESCRIPTION=React
 FAVICON=favicon.ico
+BASEURL=/
