+++ .gitignore
@@ -1,4 +1,5 @@
-cache
-coverage
-lib
 node_modules
+lib
+build
+coverage
+cache
