+++ .jestrc.json
@@ -4,77 +4,52 @@
   "cacheDirectory": "cache",
   "clearMocks": true,
   "collectCoverage": true,
-  "collectCoverageFrom": ["src/**/*.tsx", "src/**/*.ts"],
+  "collectCoverageFrom": ["src/**/*.ts", "src/**/*.tsx"],
   "coverageDirectory": "coverage",
   "coveragePathIgnorePatterns": ["node_modules", "lib"],
   "coverageProvider": "v8",
   "coverageReporters": ["json", "text", "lcov", "clover"],
-  "coverageThreshold": {
-    "global": {
-      "branches": 100,
-      "functions": 100,
-      "lines": 100,
-      "statements": 0
-    }
-  },
+  // "coverageThreshold": undefined,
   // "dependencyExtractor": undefined,
-  // "displayName": undefined,
   "errorOnDeprecated": false,
-  "extensionsToTreatAsEsm": [".tsx", ".ts"],
   "fakeTimers": {
     "enableGlobally": false
   },
   "forceCoverageMatch": [],
-  "globals": {},
   // "globalSetup": undefined,
   // "globalTeardown": undefined,
-  // "haste": undefined,
-  "maxConcurrency": 10,
+  "globals": {},
   "maxWorkers": "50%",
   "moduleDirectories": ["node_modules", "src"],
-  "moduleFileExtensions": ["tsx", "ts", "js", "json"],
+  "moduleFileExtensions": ["js", "ts", "tsx", "json"],
   "moduleNameMapper": {},
   "modulePathIgnorePatterns": [],
-  "modulePaths": [],
   "notify": false,
   "notifyMode": "failure-change",
-  "openHandlesTimeout": 1000,
   "preset": "ts-jest",
-  "prettierPath": "prettier",
-  "randomize": false,
   // "projects": undefined,
   // "reporters": undefined,
   "resetMocks": false,
   "resetModules": false,
   // "resolver": undefined,
   "restoreMocks": false,
-  "rootDir": ".",
+  // "rootDir": undefined,
   "roots": ["<rootDir>"],
   "runner": "jest-runner",
-  // "sandboxInjectedGlobals": undefined,
   "setupFiles": [],
   "setupFilesAfterEnv": [],
-  "showSeed": false,
-  "slowTestThreshold": 2000,
-  "snapshotFormat": {
-    "escapeString": false,
-    "printBasicPrototype": false
-  },
-  // "snapshotResolver": undefined,
+  "slowTestThreshold": 5,
   "snapshotSerializers": [],
   "testEnvironment": "jsdom",
   "testEnvironmentOptions": {},
-  "testFailureExitCode": 1,
   "testLocationInResults": false,
-  "testMatch": ["<rootDir>/test/**/*.tsx", "<rootDir>/test/**/*.ts"],
+  "testMatch": ["<rootDir>/test/**/*.ts"],
   "testPathIgnorePatterns": ["node_modules"],
   "testRegex": [],
   // "testResultsProcessor": undefined,
   "testRunner": "jest-circus/runner",
-  "testSequencer": "@jest/test-sequencer",
-  "testTimeout": 2,
   "transform": {
-    "^.+\\.tsx?$": [
+    "^.+\\.ts$": [
       "ts-jest",
       {
         "tsconfig": ".jesttypescriptrc.json"
@@ -85,7 +60,5 @@
   // "unmockedModulePathPatterns": undefined,
   // "verbose": undefined,
   "watchPathIgnorePatterns": [],
-  "watchPlugins": [],
-  "watchman": true,
-  "workerThreads": false
+  "watchman": true
 }
