+++ .typescriptrc.json
@@ -2,98 +2,98 @@
   // "extends": "",
   "references": [],
   // "files": [],
-  "include": ["src/**/*.tsx", "src/**/*.ts"],
+  "include": ["src/**/*.ts", "src/**/*.tsx"],
   "exclude": [],
   "compilerOptions": {
-    "allowArbitraryExtensions": true,
-    "allowImportingTsExtensions": false,
-    "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "allowUmdGlobalAccess": true,
-    "allowUnreachableCode": true,
-    "allowUnusedLabels": true,
-    "alwaysStrict": true,
-    "baseUrl": "src",
-    "checkJs": false,
-    "composite": true,
-    "customConditions": [],
-    "declaration": true,
-    "declarationDir": "lib",
-    "declarationMap": true,
-    "disableReferencedProjectLoad": true,
-    "disableSolutionSearching": true,
-    "disableSourceOfProjectReferenceRedirect": true,
-    "downlevelIteration": true,
-    "emitBOM": true,
-    "emitDeclarationOnly": false,
-    "emitDecoratorMetadata": true,
-    "esModuleInterop": true,
-    "exactOptionalPropertyTypes": true,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "importsNotUsedAsValues": "remove",
     "incremental": true,
-    "inlineSourceMap": false,
-    "inlineSources": true,
-    "isolatedModules": true,
+    "composite": true,
+    "tsBuildInfoFile": "cache/build",
+    "disableSourceOfProjectReferenceRedirect": true,
+    "disableSolutionSearching": true,
+    "disableReferencedProjectLoad": true,
+    "target": "es2022",
+    "lib": ["es2022", "dom"],
     "jsx": "react",
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
     "jsxFactory": "",
     "jsxFragmentFactory": "",
     "jsxImportSource": "",
-    "lib": ["es2022", "dom"],
-    "mapRoot": ".",
-    "maxNodeModuleJsDepth": 1,
-    "module": "es2022",
-    "moduleDetection": "auto",
-    "moduleResolution": "bundler",
-    "moduleSuffixes": [],
-    "newLine": "lf",
-    "noEmit": false,
-    "noEmitHelpers": true,
-    "noEmitOnError": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
+    "reactNamespace": "",
     "noLib": false,
-    "noPropertyAccessFromIndexSignature": true,
-    "noResolve": true,
-    "noUncheckedIndexedAccess": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "outDir": "lib",
-    // "outFile": "",
+    "useDefineForClassFields": true,
+    "moduleDetection": "auto",
+    "module": "es2022",
+    "rootDir": "src",
+    "moduleResolution": "bundler",
+    "baseUrl": "src",
     "paths": {
       "@source/*": ["./*"]
     },
-    "preserveConstEnums": true,
-    "preserveSymlinks": true,
-    "preserveValueImports": false,
-    "reactNamespace": "",
-    "removeComments": true,
-    "resolveJsonModule": true,
+    "rootDirs": ["src"],
+    "typeRoots": ["node_modules/@types"],
+    // "types": [""],
+    "allowUmdGlobalAccess": true,
+    "moduleSuffixes": [],
+    "allowImportingTsExtensions": false,
     "resolvePackageJsonExports": true,
     "resolvePackageJsonImports": true,
-    "rootDir": "src",
-    "rootDirs": ["src"],
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true,
+    "customConditions": [],
+    "resolveJsonModule": true,
+    "allowArbitraryExtensions": true,
+    "noResolve": true,
+    "allowJs": false,
+    "checkJs": false,
+    "maxNodeModuleJsDepth": 1,
+    "declaration": true,
+    "declarationMap": true,
+    "emitDeclarationOnly": false,
     "sourceMap": true,
+    "inlineSourceMap": false,
+    // "outFile": "",
+    "outDir": "lib",
+    "removeComments": true,
+    "noEmit": false,
+    "importHelpers": true,
+    "importsNotUsedAsValues": "remove",
+    "downlevelIteration": true,
     "sourceRoot": ".",
-    "strict": true,
-    "strictBindCallApply": true,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
+    "mapRoot": ".",
+    "inlineSources": true,
+    "emitBOM": true,
+    "newLine": "lf",
     "stripInternal": true,
-    "target": "es2022",
-    "tsBuildInfoFile": "cache/build",
-    "typeRoots": ["node_modules/@types"],
-    "types": ["node", "jest", "react", "react-dom", "react-dom/client"],
-    "useDefineForClassFields": true,
+    "noEmitHelpers": true,
+    "noEmitOnError": true,
+    "preserveConstEnums": true,
+    "declarationDir": "lib",
+    "preserveValueImports": false,
+    "isolatedModules": true,
+    "verbatimModuleSyntax": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "preserveSymlinks": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "noImplicitAny": true,
+    "strictNullChecks": true,
+    "strictFunctionTypes": true,
+    "strictBindCallApply": true,
+    "strictPropertyInitialization": true,
+    "noImplicitThis": true,
     "useUnknownInCatchVariables": true,
-    "verbatimModuleSyntax": true
+    "alwaysStrict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "exactOptionalPropertyTypes": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "noPropertyAccessFromIndexSignature": true,
+    "allowUnusedLabels": true,
+    "allowUnreachableCode": true,
+    "skipDefaultLibCheck": true,
+    "skipLibCheck": true
   }
 }
