+++ citation.cff
@@ -1,9 +1,9 @@
 cff-version: 1.2.0
-message: 'If you use this software you can cite it.'
+message: "If you use this software you can cite it."
 authors:
-  - given-names: 'Xi'
-    family-names: 'Matriarx'
-title: 'React'
+- given-names: "Xi"
+  family-names: "Matriarx"
+title: "React"
 version: 0.0.1
 date-released: 2024-02-04
-url: 'https://github.com/XiMatriarx/react'
+url: "https://github.com/XiMatriarx/react"
