+++ src/app.tsx
@@ -1,7 +1,5 @@
 import React from 'react'
 
-import './theme/index.scss'
-
 const App = () => {
   return <div id='app'></div>
 }
