+++ src/index.html
@@ -1,11 +1,11 @@
 <!doctype html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <title>React</title>
-    <noscript>JavaScript is required.</noscript>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+	<head>
+		<meta charset="utf-8">
+		<title>React</title>
+		<noscript>JavaScript is required.</noscript>
+	</head>
+	<body>
+		<div id='root'></div>
+	</body>
 </html>
