+++ test/index.ts
@@ -1,0 +1,7 @@
+import {greeting} from '../src/index.ts'
+
+describe('Index', () => {
+  test('Greeting', () => {
+    expect(greeting).toBe('Hello World!')
+  })
+})
